Start at moving likelihood to separate module. Needs more cleanup/refactor.
